feat: Refactor Empathic Building integration with new EBPusherService
- Introduced EBPusherService to handle Pusher connections and events.
- Updated EmpathicBuildingService to focus on API authentication and token management.
- Modified configuration handling to separate API and Pusher configurations.
- Enhanced event handling to utilize the new EBPusherService for event subscriptions.
- Improved error handling and logging for connection and subscription events.
- Removed deprecated methods and cleaned up the codebase for better maintainability.
