+++ modules/ingress/eb_subscriber/.env.example
@@ -74,18 +74,7 @@ SERVICE_OUTPUT_SCHEMA_ID=
 # -----------------------------------------------------------------------------
 # Empathic Building Configuration
 # -----------------------------------------------------------------------------
-# Empathic Building API base URL
-EB_BASE_URL=https://eu-api.empathicbuilding.com
-
-# Pusher configuration
-EB_PUSHER_KEY=your-key-here
-EB_PUSHER_CLUSTER=eu
-
-# Authentication: Either provide bearer token OR username/password
-# Bearer token authentication (preferred)
-# EB_BEARER_TOKEN=your-bearer-token
-
-# OR username/password authentication
+# Authentication: username/password
 EB_USERNAME=your-email@example.com
 EB_PASSWORD=your-password
 
@@ -100,12 +89,16 @@ EB_LOCATION_IDS=7
 # Example: {"7":"campus-1","8":"campus-2"}
 EB_LOCATION_TO_CAMPUS_MAP='{"7":"MP"}'
 
-# Subscribe to notifications
-EB_SUBSCRIBE_NOTIFICATIONS=true
+# Empathic Building API base URL
+# EB_BASE_URL=https://eu-api.empathicbuilding.com
+
+# Pusher configuration
+# EB_PUSHER_KEY=33d6c4f799c274f7e0bc
+# EB_PUSHER_CLUSTER=eu
 
 # Reconnection settings
-EB_RECONNECT_DELAY_MS=5000
-EB_MAX_RECONNECT_ATTEMPTS=10
+# EB_RECONNECT_DELAY_MS=5000
+# EB_MAX_RECONNECT_ATTEMPTS=10
 
 # -----------------------------------------------------------------------------
 # Script Configuration (for fetch-sensors.ts)
