+++ modules/ingress/eb_subscriber/src/deps.ts
@@ -8,6 +8,7 @@ import * as configs from "./utils/config";
 // Initialize services
 import { KafkaProducer, ServiceSchemaManager } from "@spine/messaging";
 import { EmpathicBuildingService } from "./services/EmpathicBuildingService";
+import { EBPusherService } from "./services/EBPusherService";
 import { getEmpathicBuildingConfig } from "./utils/config";
 import { ExcelService } from "./services/ExcelService";
 
@@ -21,10 +22,23 @@ const schemaManager =
     configs.SEND_TO === "kafka"
         ? new ServiceSchemaManager(configs.getSchemaRegistryConfig())
         : null;
-// Initialize Empathic Building service
-const empathicBuildingService = new EmpathicBuildingService(
-    getEmpathicBuildingConfig()
-);
+// Initialize Empathic Building services
+const { api: ebApiConfig, pusher: ebPusherConfig } =
+    getEmpathicBuildingConfig();
+const ebAPIService = new EmpathicBuildingService(ebApiConfig);
+
+// Auth provider adapter for pusher service
+const ebAuthProvider = {
+    getToken: async (): Promise<string> => {
+        // Ensure we have a valid token (authenticate will refresh if needed)
+        await ebAPIService.authenticate();
+        const token = ebAPIService.getAccessToken();
+        if (!token) throw new Error("Failed to obtain EB access token");
+        return token;
+    }
+};
+
+const ebPusherService = new EBPusherService(ebPusherConfig, ebAuthProvider);
 
 // Export dependencies
 export {
@@ -34,7 +48,8 @@ export {
     excelService,
     kafkaProducer,
     schemaManager,
-    empathicBuildingService
+    ebAPIService,
+    ebPusherService
 };
 
 export type { FastifyPluginAsync, FastifyInstance };
