+++ modules/ingress/eb_subscriber/src/index.ts
@@ -4,7 +4,7 @@ import {
     configs,
     kafkaProducer,
     schemaManager,
-    empathicBuildingService
+    ebPusherService
 } from "./deps";
 import { logger } from "@spine/shared";
 import { healthRoutes } from "./routes/health";
@@ -66,22 +66,33 @@ async function main() {
         logger.info("Schema manager initialized successfully");
     }
 
+    // Empathic Building Pusher service events (logging only - service handles reconnect)
+    ebPusherService.on("connected", () => {
+        logger.info("Empathic Building Pusher: connected");
+    });
+
+    ebPusherService.on("disconnected", () => {
+        logger.warn("Empathic Building Pusher: disconnected");
+    });
+
+    ebPusherService.on("connectionError", (error: unknown) => {
+        logger.error("Empathic Building Pusher: connection error", error);
+    });
+
+    ebPusherService.on("subscriptionError", (payload: unknown) => {
+        logger.warn("Empathic Building Pusher: subscription error", payload);
+    });
+
+    // Start pusher service (it will request tokens and manage reconnects itself)
+    try {
+        await ebPusherService.connect();
+        logger.info("Empathic Building Pusher: started");
+    } catch (err) {
+        logger.warn("Empathic Building Pusher: initial connect failed", err);
+    }
+
     // Set up Empathic Building event handlers
     setupEmpathicBuildingHandlers();
-
-    // Initialize Empathic Building service
-    try {
-        logger.info("Empathic Building service: Initializing...");
-        await empathicBuildingService.connect();
-        logger.info("Empathic Building service: Initialized successfully");
-        const status = empathicBuildingService.getStatus();
-        logger.info("Empathic Building service status:", status);
-    } catch (error) {
-        logger.warn(
-            `Empathic Building service failed to initialize, continuing without it:`,
-            error
-        );
-    }
 
     // Handle graceful shutdown
     const shutdown = async () => {
@@ -89,7 +100,7 @@ async function main() {
 
         // Disconnect from Empathic Building service
         try {
-            await empathicBuildingService.disconnect();
+            await ebPusherService.disconnect();
         } catch (error) {
             logger.error(
                 "Error disconnecting from Empathic Building service:",
