+++ modules/ingress/eb_subscriber/src/utils/config/eb_config.ts
@@ -1,63 +1,3 @@
-type SupportedMethod = "GET" | "POST";
-type AuthType = "none" | "basic" | "bearer" | "apikey" | "oauth2";
-type ApiKeyLocation = "header" | "query";
-type PaginationMode = "none" | "page" | "cursor" | "link";
-
-interface RestAuthConfig {
-    type: AuthType;
-    username?: string;
-    password?: string;
-    bearerToken?: string;
-    apiKey?: string;
-    apiKeyHeader?: string;
-    apiKeyQueryParam?: string;
-    apiKeyLocation?: ApiKeyLocation;
-    customHeaders?: Record<string, string>;
-    oauth?: {
-        tokenUrl: string;
-        clientId: string;
-        clientSecret: string;
-        scope?: string;
-        audience?: string;
-        grantType: string;
-        refreshMarginSeconds: number;
-    };
-}
-
-interface RestEndpointConfig {
-    path: string;
-    method: SupportedMethod;
-    bodyTemplate?: unknown;
-}
-
-interface RestPaginationConfig {
-    mode: PaginationMode;
-    pageParam?: string;
-    pageSizeParam?: string;
-    pageSize?: number;
-    maxPages?: number;
-    cursorParam?: string;
-    nextCursorField?: string;
-    nextLinkField?: string;
-}
-
-interface RestPollingConfig {
-    pollIntervalMs: number;
-    timeoutMs: number;
-    retryAttempts: number;
-    retryDelayMs: number;
-}
-
-interface RestApiConfig {
-    baseUrl: string;
-    endpoints: RestEndpointConfig[];
-    poller: RestPollingConfig;
-    auth: RestAuthConfig;
-    pagination: RestPaginationConfig;
-    customHeaders: Record<string, string>;
-    defaultMethod: SupportedMethod;
-}
-
 const parseNumber = (value: string | undefined, fallback: number): number => {
     if (!value) {
         return fallback;
@@ -78,65 +18,67 @@ const parseJSON = (value: string | undefined) => {
 };
 
 // Empathic Building API configuration
-import type { EmpathicBuildingConfig } from "../eb_types";
+import type { EBApiConfig, EBPusherConfig } from "../eb_types";
 
-let empathicBuildingConfig: EmpathicBuildingConfig | undefined = undefined;
+let ebApiConfig: EBApiConfig | undefined = undefined;
+let ebPusherConfig: EBPusherConfig | undefined = undefined;
 
-const getEmpathicBuildingConfig = (): EmpathicBuildingConfig => {
-    if (empathicBuildingConfig) {
-        return empathicBuildingConfig;
+const getEmpathicBuildingConfig = (): {
+    api: EBApiConfig;
+    pusher: EBPusherConfig;
+} => {
+    if (ebApiConfig && ebPusherConfig) {
+        return { api: ebApiConfig, pusher: ebPusherConfig };
     }
 
-    const EB_BASE_URL =
-        process.env.EB_BASE_URL || "https://eu-api.empathicbuilding.com";
-    const EB_PUSHER_KEY = process.env.EB_PUSHER_KEY || "33d6c4f799c274f7e0bc";
-    const EB_PUSHER_CLUSTER = process.env.EB_PUSHER_CLUSTER || "eu";
-    const EB_BEARER_TOKEN = process.env.EB_BEARER_TOKEN;
+    // Required configurations with no defaults
     const EB_USERNAME = process.env.EB_USERNAME;
     const EB_PASSWORD = process.env.EB_PASSWORD;
     const EB_ORGANIZATION_IDS = process.env.EB_ORGANIZATION_IDS;
     const EB_LOCATION_IDS = process.env.EB_LOCATION_IDS;
-    const EB_SUBSCRIBE_NOTIFICATIONS =
-        process.env.EB_SUBSCRIBE_NOTIFICATIONS === "true";
-    const EB_RECONNECT_DELAY_MS = process.env.EB_RECONNECT_DELAY_MS;
-    const EB_MAX_RECONNECT_ATTEMPTS = process.env.EB_MAX_RECONNECT_ATTEMPTS;
 
     // Validate authentication: either bearerToken or username/password
-    if (!EB_BEARER_TOKEN && (!EB_USERNAME || !EB_PASSWORD)) {
+    if (!EB_USERNAME || !EB_PASSWORD) {
         throw new Error(
             `Empathic Building configuration is not set: Either EB_BEARER_TOKEN or both EB_USERNAME and EB_PASSWORD must be provided`
         );
     }
 
-    if (
-        !EB_ORGANIZATION_IDS &&
-        !EB_LOCATION_IDS &&
-        !EB_SUBSCRIBE_NOTIFICATIONS
-    ) {
+    if (!EB_ORGANIZATION_IDS && !EB_LOCATION_IDS) {
         throw new Error(
             `Empathic Building configuration is not set: At least one of EB_ORGANIZATION_IDS, EB_LOCATION_IDS, or EB_SUBSCRIBE_NOTIFICATIONS must be configured`
         );
     }
 
-    empathicBuildingConfig = {
+    // Optional configurations with defaults
+    const EB_BASE_URL =
+        process.env.EB_BASE_URL || "https://eu-api.empathicbuilding.com";
+    const EB_PUSHER_KEY = process.env.EB_PUSHER_KEY || "33d6c4f799c274f7e0bc";
+    const EB_PUSHER_CLUSTER = process.env.EB_PUSHER_CLUSTER || "eu";
+    const EB_RECONNECT_DELAY_MS = process.env.EB_RECONNECT_DELAY_MS;
+    const EB_MAX_RECONNECT_ATTEMPTS = process.env.EB_MAX_RECONNECT_ATTEMPTS;
+
+    ebApiConfig = {
+        baseUrl: EB_BASE_URL,
+        username: EB_USERNAME,
+        password: EB_PASSWORD
+    };
+
+    ebPusherConfig = {
         baseUrl: EB_BASE_URL,
         pusherKey: EB_PUSHER_KEY,
         pusherCluster: EB_PUSHER_CLUSTER,
-        bearerToken: EB_BEARER_TOKEN,
-        username: EB_USERNAME,
-        password: EB_PASSWORD,
         organizationIds: EB_ORGANIZATION_IDS
             ? EB_ORGANIZATION_IDS.split(",").map((id) => id.trim())
-            : undefined,
+            : [],
         locationIds: EB_LOCATION_IDS
             ? EB_LOCATION_IDS.split(",").map((id) => id.trim())
-            : undefined,
-        subscribeToNotifications: EB_SUBSCRIBE_NOTIFICATIONS,
+            : [],
         reconnectDelayMs: parseNumber(EB_RECONNECT_DELAY_MS, 5000),
         maxReconnectAttempts: parseNumber(EB_MAX_RECONNECT_ATTEMPTS, 10)
     };
 
-    return empathicBuildingConfig;
+    return { api: ebApiConfig, pusher: ebPusherConfig };
 };
 
 let locationToCampusMap: Map<string, string> | undefined = undefined;
@@ -172,12 +114,4 @@ const getLocationToCampusMap = (): Map<string, string> => {
     return locationToCampusMap;
 };
 
-export {
-    getEmpathicBuildingConfig,
-    getLocationToCampusMap,
-    type RestApiConfig,
-    type RestEndpointConfig,
-    type RestPaginationConfig,
-    type RestPollingConfig,
-    type RestAuthConfig
-};
+export { getEmpathicBuildingConfig, getLocationToCampusMap };
